Restart gunicorn instead of reloading it
Because we intend to launch a completely different executable
we should stop the old process and bring up a new one instead
of telling the existing one to reload.

This solves a problem where upstart would lose track of the
current PID and where gunicorn would exit during start up due
to an existing PID file.