Kill dotted view syntax for url reversing
Per django 1.8 docs, the 'app.views.viewname' method of url reversing
in templates with the url tag is deprecated and will be removed in
django 1.10.

This patchset creates a consistent set of names for all urls that need
to be reversed. All names are prefixed with app_name, for example,
lifestream-index.

This patchset also renames the existing lifestream view functions so
they are not prefixed with app_name. This brings them more in line with
other apps in hyperbola.

Fixes GH-31.

https://docs.djangoproject.com/en/1.8/ref/templates/builtins/#url