Reconstruct the original hyperbola page
The beginnings of this repository look like the assets directory for the
first version of hyperbo.la. It houses jQuery and some plugins as well as
js and css that appears to be written by me (it even has the hyperbola
link color!)

All that was missing was an index.html.

By browsing the js and css I was able to reconstruct the html layout of
the page and get it to render, crappy jQuery fade animations and all.