Fix media backup
Since the move to hyperbola3, the layout of the media file system path
has changed from being /hyperbola/media to a env-specific symlink. This
caused corrupted, empty tarballs to be backed up.

To fix, supply the dereference arg to tarfile.open to make it follow
symlinks. The path prefix also changes from media to the name of the
env-specific symlink.

Remove colon from backup names for portability.

This work brought to you by National Check Your Backups Day and Gitlab.