Create hyperboladc.net and hyperbolausercontent.net
These domains were bought with route53 domains and imported to
terraform with terraforming.

A hosted zone is created in hyperboladc.net for each datacenter.
These zones are created via the hyperbola/dc module.

Current zones are:
- aws
- linode
- local

Add A records for hyperbola3 and bastion.
Update vagrant ansible config for wiki to point to new hyperboladc.net
domain.