Generate text outline from 'yperbola' text and embed in logo SVG
This removes the dependency on fonts which are not present on iOS. This fixes
a regression introduced when migrating to webpack. Previously, PNGs were generated
for the logo which stored a rasterized version of the text.

This is a cherry-pick of a35826f from the maintentance branch
hyperbola3-maintenance. This commit differs in the generated css
bundle and some fixes to get yperbola-outline.js to pass eslint.