Enable turning off NAT in network terraform module
NAT represents 45% of my AWS bill. In November, 195kb of traffic
traversed it. hyperbola-app servers do not talk to the public
internet since the servers are immutable and S3 is accessed via
a VPC endpoint.

hyperbola-app only uses the network to send backups to writing
user content to S3, reading resumes from S3, and talking to the
ALB.

This commit exercised bugs in the terraform configuration when the
NAT was not enabled causing route tables to disappear